requirements.txt
tracdap-runtime == 0.5.30
pandas~=1.5.0
numpy~=1.23.3
scipy==1.11.1